Turn off enforceFocus option for react-bootstrap-modal
Right now, the react-bootstrap-modal code we use runs into issues with
selectro. Selectro relies on observing and changing user focus to
function. However, the react-bootstrap-modal code uses a  react-overlay
that has an `enforceFocus` option turned on by default, which causes the
modal to grab back focus if it is ever lost. This leads to focus
contention between the modal and selectro, preventing the selectro from
being searchable, and also causing the selectro component to constantly
re-fetch records.